+++ concordium-std-derive/src/lib.rs
@@ -1559,19 +1559,27 @@ const RESERVED_ERROR_CODES: i32 = i32::MIN + 100;
 /// Creating custom enums for error types can provide meaningful error messages
 /// to the user of the smart contract.
 ///
-/// Note that at the moment, we can only derive fieldless enums and the return
-/// value is always [`None`][std::option::Option::None].
+/// For enums where at least one variant has a field, the enum *must* implement
+/// [`Serial`](../concordium_contracts_common/trait.Serial.html) if [`Reject`]
+/// is to be derived.
+///
+/// For variants with fields, the enum is serialized into the buffer used in
+/// the return value, which is always [`Some`][std::option::Option::Some].
+/// The contents of the buffer are returned by the contract along with the error
+/// code.
+/// For fieldless variants the return value is always
+/// [`None`][std::option::Option::None] and only the error code is returned.
 ///
 /// The conversion will map the first variant to error code -1, second to -2,
 /// etc.
 ///
 /// ### Example
 /// ```ignore
-/// #[derive(Clone, Copy, Reject)]
+/// #[derive(Reject, Serial)]
 /// enum MyError {
 ///     IllegalState, // receives error code -1
 ///     WrongSender, // receives error code -2
-///     // TimeExpired(time: Timestamp), /* currently not supported */
+///     TimeExpired(time: Timestamp), // receives error code -3 (enums with fields must implement Serial)
 ///     ...
 /// }
 /// ```
@@ -1608,6 +1616,8 @@ fn reject_derive_worker(input: TokenStream) -> syn::Result<TokenStream> {
     };
 
     let variant_error_conversions = generate_variant_error_conversions(enum_data, enum_ident)?;
+    let enum_var_ident = format_ident!("{}", "e");
+    let variant_matches = generate_variant_matches(enum_data, enum_ident, &enum_var_ident);
 
     let gen = quote! {
         /// The from implementation maps the first variant to -1, second to -2, etc.
@@ -1616,10 +1626,9 @@ fn reject_derive_worker(input: TokenStream) -> syn::Result<TokenStream> {
         #[automatically_derived]
         impl From<#enum_ident> for Reject {
             #[inline(always)]
-            fn from(e: #enum_ident) -> Self {
-                Reject {
-                    error_code: unsafe { concordium_std::num::NonZeroI32::new_unchecked(-(e as i32) - 1) },
-                    return_value: None
+            fn from(#enum_var_ident: #enum_ident) -> Self {
+                match &#enum_var_ident {
+                   #variant_matches
                 }
             }
         }
@@ -1627,6 +1636,57 @@ fn reject_derive_worker(input: TokenStream) -> syn::Result<TokenStream> {
         #(#variant_error_conversions)*
     };
     Ok(gen.into())
+}
+
+/// Generate the cases for matching on the enum. For variants with fields, the
+/// whole enum type is serialized into a buffer that is included in the return
+/// value. This means that enums with fields need to implement `Serial` in order
+/// for deriving `Reject` to work. For fieldless variants, the return value is
+/// always `None`.
+fn generate_variant_matches(
+    enum_data: &DataEnum,
+    enum_name: &syn::Ident,
+    enum_var_ident: &syn::Ident,
+) -> proc_macro2::TokenStream {
+    let mut match_cases = proc_macro2::TokenStream::new();
+    for (index, variant) in enum_data.variants.iter().enumerate() {
+        let variant_ident = &variant.ident;
+        match variant.fields {
+            syn::Fields::Named(_) => {
+                match_cases.extend(quote! {
+                    #enum_name::#variant_ident{..} => Reject {
+                        error_code: unsafe { num::NonZeroI32::new_unchecked(-(#index as i32) - 1) },
+                        return_value: {
+                            let mut buf = Vec::new();
+                            concordium_std::Serial::serial(&#enum_var_ident, &mut buf).unwrap_abort();
+                            Some(buf)
+                        },
+                    },
+                });
+            }
+            syn::Fields::Unnamed(_) => {
+                match_cases.extend(quote! {
+                    #enum_name::#variant_ident(..) => Reject {
+                        error_code: unsafe { num::NonZeroI32::new_unchecked(-(#index as i32) - 1) },
+                        return_value: {
+                            let mut buf = Vec::new();
+                            concordium_std::Serial::serial(&#enum_var_ident, &mut buf).unwrap_abort();
+                            Some(buf)
+                        },
+                    },
+                });
+            }
+            syn::Fields::Unit => {
+                match_cases.extend(quote! {
+                    #enum_name::#variant_ident => Reject {
+                        error_code: unsafe { num::NonZeroI32::new_unchecked(-(#index as i32) - 1) },
+                        return_value: None,
+                    },
+                });
+            }
+        };
+    }
+    match_cases
 }
 
 /// Generate error conversions for enum variants e.g. for converting